+++ NEWS.md
@@ -1,6 +1,8 @@
+
+
+ - add is_monotonic and is_strictly_monotonic
+
 # monotone 0.3.0
 
  - Added a `NEWS.md` file to track changes to the package.
  - Fixed documentation
-
-